Refactor: Add servicelayer
